+++ packages/stentor-determiner/README.md
@@ -1,0 +1,3 @@
+## stentor-context
+
+Context for stentor
