+++ packages/stentor-determiner/package.json
@@ -1,0 +1,44 @@
+{
+  "name": "stentor-determiner",
+  "license": "Apache-2.0",
+  "publishConfig": {
+    "access": "public"
+  },
+  "version": "1.23.34",
+  "description": "Determination for stentor",
+  "types": "lib/index",
+  "main": "lib/index",
+  "files": [
+    "lib"
+  ],
+  "engines": {
+    "node": "10.x"
+  },
+  "devDependencies": {
+    "@types/chai": "4.2.11",
+    "@types/sinon": "7.5.2",
+    "@types/sinon-chai": "3.2.3",
+    "@xapp/config": "0.2.3",
+    "chai": "4.2.0",
+    "mocha": "7.1.1",
+    "sinon": "9.0.1",
+    "sinon-chai": "3.5.0",
+    "stentor-models": "1.23.34",
+    "ts-node": "8.2.0",
+    "typescript": "3.8.3"
+  },
+  "dependencies": {
+    "stentor-request":"1.23.34",
+    "stentor-storage": "1.23.34",
+    "stentor-time":"1.23.34",
+    "stentor-utils":"1.23.34"
+  },
+  "peerDependencies": {
+    "stentor-models": "1.x"
+  },
+  "scripts": {
+    "build": "tsc -d true -p .",
+    "clean": "rm -rf ./lib/*",
+    "test": "mocha --recursive -r ts-node/register \"./src/**/*.test.ts\""
+  }
+}
