+++ packages/stentor-determiner/src/Guards.ts
@@ -1,5 +1,5 @@
 /*! Copyright (c) 2019, XAPPmedia */
-import { Conditional, Contextual, JSONDependable, JSONDependent } from "stentor-models";
+import { Conditional, JSONDependable, JSONDependent } from "stentor-models";
 
 /**
  * Guard to determine if the object implements JSONDependable.
@@ -11,10 +11,6 @@ export function isJSONDependable<T>(item: object): item is JSONDependable<T> {
     return !!item && !!(item as JSONDependent).JSONPathMatch;
 }
 
-export function isContextual(item: object): item is Contextual {
-    return !!item && typeof (item as Contextual).context === "object";
-}
-
 /**
  * Type guard to check if the object implements Conditional
  * 
@@ -24,3 +20,4 @@ export function isContextual(item: object): item is Contextual {
 export function isConditional<T>(item: object): item is Conditional<T> {
     return !!item && (typeof (item as Conditional<T>).conditions === "object" || typeof (item as Conditional<T>).conditions === "string");
 }
+
