+++ packages/stentor-determiner/src/compileJSONPaths.ts
