+++ packages/stentor-determiner/src/determine.ts
