+++ packages/stentor-guards/README.md
@@ -1,0 +1,3 @@
+## stentor-response
+
+Response for :mega: stentor
