+++ packages/stentor-guards/src/Guards.ts
@@ -12,25 +12,6 @@ import {
     TimeContextual
 } from "stentor-models";
 
-/**
- * Determines if the item has a schedule.
- */
-export function isScheduled<T>(item: object): item is Scheduled<T> {
-    return !!item && (item as Schedulable).schedule !== undefined;
-}
-
-export function isActiveWithin<T>(item: object): item is ActiveWithin<T> {
-    return !!item && (item as ActiveWithinable).activeWithin !== undefined;
-}
-
-export function isFirstTime<T>(item: object): item is FirstTime<T> {
-    return !!item && (item as FirstTimeable).firstTime !== undefined;
-}
-
-export function isHaveNotSeenWithin<T>(item: object): item is HaveNotSeenWithin<T> {
-    return !!item && (item as HaveNotSeenWithinable).haveNotSeenWithin !== undefined;
-}
-
 export function isLastActive<T>(item: object): item is LastActive<T> {
     return isActiveWithin(item) || isFirstTime(item) || isHaveNotSeenWithin(item);
 }
