+++ packages/stentor-guards/src/isActionable.ts
