+++ packages/stentor-guards/src/isContextual.ts
@@ -1,0 +1,6 @@
+/*! Copyright (c) 2019, XAPPmedia */
+import { Contextual } from "stentor-models";
+
+export function isContextual(item: object): item is Contextual {
+    return !!item && typeof (item as Contextual).context === "object";
+}
