+++ packages/stentor-guards/src/isIntent.ts
@@ -1,15 +1,8 @@
 /*! Copyright (c) 2019, XAPPmedia */
-import { Handler, Intent } from "stentor-models";
+import { Intent, Handler } from "stentor-models";
 import { existsAndNotEmpty } from "stentor-utils";
+import { isHandler } from "./isHandler";
 
-/**
- * Determine if the request handler props are for a handler.
- *
- * @public
- */
-export function isHandler(props: Handler | Intent): props is Handler {
-    return !!props && (props as Handler).type !== undefined && (props as Handler).type !== null;
-}
 
 /**
  * Determine if the props are for an Intent
