+++ packages/stentor-guards/src/isLastActiveResponse.ts
@@ -13,13 +13,3 @@ export function isLastActiveResponse(response: Response): response is LastActive
     return !!response && isLastActive(response);
 }
 
-/**
- * Type guard to determine if the Response is a SchedulableResponse
- *
- * @export
- * @param {Response} response
- * @returns {response is SchedulableResponse}
- */
-export function isSchedulableResponse(response: Response): response is SchedulableResponse {
-    return !!response && isScheduled(response);
-}
