+++ packages/stentor-guards/src/isLinkoutSuggestion.ts
