+++ packages/stentor-guards/tsconfig.json
@@ -4,6 +4,6 @@
     "rootDir": "src",
     "outDir": "lib"
   },
-  "include": ["src", "../stentor-guards/src/Guards.ts"],
+  "include": ["src"],
   "exclude": ["node_modules", "src/**/__test__/*"]
 }
