--- packages/stentor-handler/src/Content/index.ts
@@ -1,3 +1,0 @@
-/*! Copyright (c) 2019, XAPPmedia */
-export * from "./Contentable";
-export * from "./Guards";
