--- packages/stentor-response/src/Suggestion/__test__/.npmignore
@@ -1,2 +1,0 @@
-# Ignore this whole directory
-*
