--- packages/stentor-response/src/Suggestion/index.ts
@@ -1,2 +1,0 @@
-/*! Copyright (c) 2019, XAPPmedia */
-export * from "./Guards";
