+++ packages/stentor-response/src/determineSegment.ts
