+++ packages/stentor-response/src/getResponse.ts
