+++ packages/stentor-response/src/index.ts
@@ -3,7 +3,7 @@ export * from "./canFulfillAll";
 export * from "./canFulfillNothing";
 export * from "./concat";
 export * from "./Display";
-export * from "./Guards";
+export * from "../../stentor-guards/src/Guards";
 export * from "./ResponseBuilder";
 export * from "./Suggestion";
 export * from "./TemplatedResponseOutput";
